ath1Xk: restrict user country setting during active vdev
Currently, if user wants, country setting can be changed on the fly after
the interface(s) have started beaconing. However, hostapd ignores this regulatory
change event since according to its state, it did not ask for it. This leads
to interfaces which were already active continue beaconing with same old data
which is wrong.

Since there is no potential usecase of changing regulatory on the fly, hence
add change to restrict changing country setting if any of the vdev in it is
active.

Also since regulatory applies to whole SOC, restrict this further to not allow
if any of pdev on a SOC is having an active vdev.

Change-Id: I4a6b30d966a4cc78a4a711a4306294551d2ec013
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>