mac80211: fix skb and sdata mismatch
Currently, skbs are getting tagged with wrong sdata, i.e even if the
sdata does not support a skb specific channel, we are tagging the sdata
along with that.

In issue scenerio, 1 DUT was having 6G Mesh(root) + 2G AP and the other DUT
was having 2G Mesh(child) , and the packets from 2G Mesh were seeing in
6G Mesh, interface.

In this fix, tagging the sdata only if the respective skb's channel support is
present in the corresponding sdata.

Change-Id: I4bd5b0037c81890bd0dcdb0e95dfe4ed521e355f
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>