mac80211.sh: fix adding freq list for STA mode
Currently, in single wiphy architecture, when sta interface is brought up
in one particular band from OpenWrt config file, it used to scan for
all the bands instead of one.

Changes were made to scan only in that particular band by adding a freq list
in the config. However the helper function was tighlty coupled with split radio
devices which is wrong.

Due to above, still in non-spit radio, scan was being done in all bands.

In max vaps case this leads to the following prints -
[ 166.141362] ath12k_pci 0002:01:00.0: failed to create vdev. No free vdev id left.
[ 166.141386] ath12k_pci 0002:01:00.0: unable to create scan vdev -22

This happens because - 
* When STA comes up, it tries to initiate scan in all bands.
* In that band if already it is operating in max vaps, and it cant re-use those
  interfaces for scanning, it will try to create a new scan interface.
* But since there is no space for new vdev, error prints are seen.

Fix this by using helper function which is agnostic to type of device and add freq
list in all cases.

Change-Id: Iba9b2cd557c96f1a5967397087909c7d5d0312f1
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>