mac80211.sh: Fix vaps not coming up after config update and reboot
Currently, when we update config and reboot the system, vaps are not coming up.
If we do a wifi up and wifi down, after reboot vaps comes up.

ML vaps were not started as single instance because mac80211_setup_interface()
is invoked directly from netifd process and the shared file was not created
based on which the decision was taken to shared in single instance hostapd or
add bss to global instance.

In case of wifi up/down wifi mac80211.sh file, pre_wifi_updown() takes care
of updating mld vap details and after reboot interfaces are up.

To resolve this issue add changes to derive radio vap details and ml_vap details
when shared file is not present.

Change-Id: I3483f7a82c88b5be69b4ec07c98d815204741b12
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>