wifi: ath12k: Address corruption in skb->mark field and invalid Queue_id sent to telemetry agent modules
When "qca-nss-nsm.ko", in not loaded which is needed for delay stats,
the current time value which is a u32 byte variable is overflowing
the skb->mark, due to which the packet is not identified as a SAWF skb
during tx completion due to which the needed stats are not
incremented. Fixed this by addressing the overflow issue.

Also handled the return value of clear stats command from vendor NL.
Added "telemetry_peer_ctx" checks before calling specific APIs.
Added support to get the configured moving average packet number
for sending the stats to telemetry_agent.

Wrong Queue_id was getting sent to telemetry agent to get specific
stats. Addressed that issue by combining the tid and q_id as the
telemetry_agent uses 1D array and our driver uses 2D array for per
peer SAWF queues.

Change-Id: I36994e16c56c40d1698fb865900fb396d5b1a8e6
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>