mac80211: Add wmm default values to individual links
Problem statement:
When hostapd sends out wmm values only to the first link
within a BSS, it is the driver's responsibility to replicate
those values to other links or set them to default based on the
access category. If we set the values only to the deflink, then
we see zero txq_params values for SLO/MLO scenario.

Solution:
Set wmm default values to individual links during update_vif_links, so
that other links within different BSS are able to have non-zero values
which does not pose a problem while configuring tx.

Change-Id: I2b5f35e0e59954d3accf3eca3a97bd042d0ad971
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>