ath12k: Skip sta stale entry deletion during FW recovery
Problem statement:
When FW is asserted and recovery takes place, usually for
a partner chip assertion scenario, the recovery time is
relatively high compared to a single chip recovery. During
this recovery time, there is a chance that the upper layers
might kickout STA entries from the driver, which FW won't
entertain as it is already down. Also, when recovery happens
the FW starts fresh, meaning it does not have any entries
and starts with adding the existing/new entries. So, in
both the situations, the deletion is not needed.

Solution:
Avoid sta stale entry deletion till recovery is complete.

Change-Id: Iaf1956ac086a58e06eb4c0e7e9de80a1bb3aa922
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>