ath12k: set link_ndev to NULL after unregister_netdev
arvif->ndev_pvt->link_ndev is dangling after the unregister_netdev. So
avoid dangling pointer, link_ndev is set to NULL.

Rename ath12k_bond_link_delete to ath12k_bond_link_release and
ath12k_bond_link_add to ath12k_bond_link_enslave.

Added a null check for link_ndev in ds_disable subroutine.

When recovery in progress and reboot is issued, there is no need to
wait for Subsystem recovery (SSR) to be completed.

When a SSR starts in the middle of wifi restart, some of the
arvifs are enslaved with bond dev however they are not fully up and
beaconing. Upper layer is not aware of the SSR recovery and the process
to bring up the interfaces are failed due to the shutdown of the chip.
After the SSR recovery is completed and wifi down is issued, then it is
possible that some of the arvifs are enslaved with bond dev and the
upper layer will not do bring down sequence. cfg80211 notifier will
call un-register process only and start un-registering the net dev. When
a link dev is un-registered before it has been released from bond dev
then netdev will throw a below warning.

"unregister_netdevice: waiting for wlan1_l1 to become free. Usage count = 1"

For this scenario, release the link ndevs for all those arvifs which are
enslaved but got interrupted due to SSR or wifi restart.

For 2G band, the enslaving and releasing a link_ndev is supported.
However, PPE-DS is disabled. Modified the compilation flag in core reset
to allow bond release for arvifs from 2G radio.

Change-Id: Id6dc367aa75bd2cfad12ce240718a614b72d7005
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>