wifi: ath12k: fix rcu stall in raw mode
When BiDi traffic is initiated in raw mode, rcu stall
is seen immediately with unresponsive console.

This is seen because of skb prefetch of same address
everytime in ath12k_dp_rx_get_msdu_last_buf() api.

Remove the prefetch changes in ath12k_dp_rx_get_msdu_last_buf()
as it is not needed when skb queue walk is done.

Fixes: Ibf6c83ab8da (wifi: ath12k: Rx data path optimizations)

Change-Id: I749be60811fe88617ba9a2a5a00adc1f43b5edf0
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>