mac80211.sh: sta fails to join in 5G high band
Currently, in RDP454 when auto channel is configured
the frequency list in wpa_supplicant has all the
frequencies except 5G high band which is not expected.

If the channel is configured as auto, only that particular
band frequency list has to be configured in wpa_supplicant.

Add changes to handle auto channel scenario in split phy.

Change-Id: I740d6201eb764e8ae25dfeeba5506e8051fb6cf6
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>