apsta_mode.sh: Repeater vap is configured incorrectly
When Multi-link ap interface (say 2 GHz + 5 GHz + 6 GHz) is
configured as repeater with sta interface (say 5 GHz).
Once the sta is associated, that particular channel is not
configured to approriate repeater interface link.

This is because, at time the sta_freq_list argument in the
script is null and this leads to incorrect link id config. i.e.,
5 GHz channel is configured to 2 GHz channel

To address this, add changes to fetch sta supported frequency
list api and from that get the link information to configure
for the ap interface

Change-Id: Ic736c96dae753c660e59866e4550f0fb0482b4eb
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>