cfg80211: DFS - Increase grace period for PRE CAC timeout to 8 s
Sometimes, the increased 4 second time is not enough for the CAC
to be finished, especially when there are two 5G worker jobs
in case of a split-phy model. If the dfs_cac_time_worker gets
scheduled post the grace period blanket, the AP does not come
up correctly on 5G high.

Change the pre_cac grace timer to 8 from 4,
since the overall operation takes close to 7s at max.

Change-Id: I4be0ffea6bae045d669ddae6c95a44cf8bea13da
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>