ath12k: Fix q6 crash caused due to parallel qmi handshake for all chips
QMI handshake for all chips are triggered in parallel due to this mlo
memory allocations are not in order. QMI handshake is expected to execute
sequentially, that is master chip should complete handshake and then
followed by other chips.
Reverted the change which sends host caps for all chips one after
the other in a loop. Now host cap is sent for a chip after receiving
server ready for previous chip.

Fixes: ath12k: I09db0839d4e(Add rproc recovery support in ath12k)

Change-Id: Ic9589cfdb06675e088bbfa721b30e8cbdbed3c88
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>