mac80211: fix dropping frame from unmatched freq of same hw
Sometimes mac80211 receives a frame from other bandwidth freq which will
not be same as the configured channel center freq, hence mac80211 drops
that frame.

Frames of same hardwares are need to be processed should not be dropped
so compare the frequencies by hw idx using cfg80211_get_hw_idx_by_freq
and give it to upper stack.

Fixes: I4bd5b0037 ("mac80211: fix skb and sdata mismatch")

Change-Id: I1b1571d605f18331c33e7b6525264fd75269635f
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>