wifi: ath12k: Fix issues observed when breach is notified to ath12k driver
When breach is detected in Telemetry agent module and the same is
notified to ath12k driver, scheduling while atomic issue is observed and
the corresponding NL event was not sent to user space.

Fixed this by unlocking mutex and including the NL event that is
needed with which the NL msg will be sent.

Change-Id: Ifb534cf029c385250f034be5a1c3c52a29fd0da6
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>