cfg80211: remove need rtnl flag for CAC start command
When radar is injected, if hostapd/any userspace
application tries to switch the channel to a
DFS channel, then start radar detection is triggered.

During this scenario, rtnl lock could not be obtained
and it results in deadlock. The deadlock scenario
occurs between radar detected and to start a cac for
new dfs channel.

This changes were brought in as part of Direct switch,
hence, temporarily disabling the need_rtnl flag for
SFE mode.

Fixes: I5ec17c390d9 (ath12k: fix RTNL locks, ACS and DFS cases for bonded interface featue)

Change-Id: Ic35156dbb1c5e175967f63c93e0202f3cae23bcd
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>