ath12k: fix ath12k_mac_op_switch_vif_chanctx() mac ops
If more than 1 link switches channel at near by same time such that CSA is
not completed on the first before it starts on later, then once the CSA is
completed on both, mac80211 sends all the channel contexts in one shot to the
driver via ath12k_mac_op_switch_vif_chanctx(). This is due to the single wiphy
architecture since all channel contexts are maintained under single wiphy.

Hence there could be a possibility that in the passed channel contexts array,
there could be contexts belonging to different underlying radios. However,
currently there is an assumption that the passed array will have channel context
belonging to a single radio and hence only 1st member of the array is used to
get the underlying radio info. This leads to sending incorrect vdev ids to
firmware as well as stack corruption since now total number of elements in array
could be more than that a radio supports.

Modify the ath12k_mac_op_switch_vif_chanctx() mac ops to get all the consecutive
contexts belonging to the same radio and further process it in a single shot.

Change-Id: I10d495e568955a2d610eb321f7dbdb96d6ff67c0
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>