wifi: ath12k: Enable UMAC reset support for DS mode
Add support for UMAC DP reset of partner SOC when
one SOC crashes in MLO as a part of MODE 1 SSR for DS Mode.

In this process shared memory will be used by Host and
Firmware to communicate the sequence of events. Host will
clear the data structures such as tx/rx descriptors and
re-initialize the rings and reset hp/tp to 0.

In addition to host handling, inform the umac reset progress
to PPE through PPE handlers so that PPE will take necessary
action to disable interrupts.

Change-Id: I67999b69d24c84713e64c33ba2dda3dc907440b6
Signed-off-by: Balamurugan S <quic_bselvara@quicinc.com>
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>