wifi: ath12k: Fix duplicate WMI Peer create command getting sent.
During peer create, currently the vdev_id is checked for duplicate
entry. This results in an issue, when two Virtual APs(VAP) are created
in same radio, and a STA is trying to associate to the first VAP and
second VAP one after another. This behaviour can occur for various
reasons like disassoc for first VAP not reaching the AP, power down
and power up of STA.

Current check condition of vdev_id will prevent the wmi peer create
command to not be sent for the VAP, which currently has that MAC
in its peer list. This peer list is specific to SOC. However, in case
of Multi VAP condition within same radio, when a STA is associated to
a VAP and a peer create request for another VAP is received with same
MAC, the firmware asserts if the peer create command is sent to a VAP,
where the radio has the same peer MAC in its list. Modified the logic
to check the pdev_id, instead of vdev_id, which implies that there
should not be two STA with same MAC connected to same radio.

Fixes: 5620d1b751ce ("wifi: ath12k: Add peer rhash table support")
Change-Id: I85d2b3718f181e1c33e781bb13ea635534f70e41
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>