mac80211.sh: fix multiple 6GHz stations connection issue
In upgraded hostapd version uses mbssid is set 1 for non ema, mbssid is set 2
for ema, but mbssid and ema is cofigured partically in hostapd, which caused
hostapd to not send rnr elems to kernel as well as not include in beacon template.
So, this causes connecting multiple 6GHz mlo stations issue in AP configured
multiple 6GHz VAPs.

To fix mbssid is set to 2(ENHANCED_MBSSID_ENABLED) in the mac80211 script
to add both mbssid and rnr in the 6G beacon.

Change-Id: I302fa9c9b7aeb74d5ea8e0b27c758dafdd414b94
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>