mac80211.sh: Fix interfaces not coming up
Currently when 2G sta and 5G+6G are in mld config interfaces are not coming up
this is because interface name is not properly assigned to the interfaces.

This issue is because we have not initialized variables sta_vaps_count and
sta_radio, this causes conditions like
"if [ $sta_vaps_count -gt 0  -o  $mld_vaps_count -gt 0 ]"
always flase if sta_vaps_count have no value. By this interface index is not
properly initiated to the interfaces that causes interface to go to disable state.

Fix this by initializing sta_vaps_count and sta_radio variables to 0.

Fixes: I30ff926c67924 ("mac80211.sh: add STA MLO support")

Change-Id: I435bbd1707e6cb1e7c4c3d6125527a8fb3dc230e
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>