ath12k: fix RU puncturing bitmap in peer channel width switch command
Currently RU puncturing bitmap for the peer is taken from the link_sta
structure's member in peer channel width switch command. However, this
member is set only for the mesh interface. Instead, peer's vif's channel
context's chandef should be used to get the value in all cases. Only if
it is a mesh peer, link_sta should be used.

This fixes the above issue.

Also, the puncturing pattern should be filled only if the peer is EHT
capable. So while at it, add EHT capability validation as well.

Fixes: 67a4ab6fa755 ("ath12k: add peer channel width switch command support")

Change-Id: If161dcb8a2af20abcefb422126218b6e095efde4
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>