ath12k: fix firmware assert seen during peer assoc command
Currently, during peer assoc, peer's maximum capable bandwidth is
sent in the peer flags where as phymode is sent as the operating
bandwidth. When sta is more capable than AP's operating bandwidth,
then we are sending phymode as AP operating bandwidth lesser than
peer_flags which is maximum bandwidth capable of STA and that leads
to firmware assert.

The expectation from firmware is to have a phy mode with an upper
limit (max value) and the corresponding peer flags with a value less
than or equal to the phy mode. (peer flags would be associated with
the current operating band)

Hence, add changes to send phymode as maximum bandwidth and peer_flags
as operating bandwidth in peer assoc command

Fixes: I5bd2ceb4e5c ("ath12k: fix firmware assert during peer channel width change")

Change-Id: Icff12459056d3cfe503be7a0e13294fed0d4ea56
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>