ath12k: reduce sta-count irrespective of recovery mode
Current station-count decrement logic is dependent
on recovery mode, and so if decrement is done when
recovery is in progress, decrement logic may fail
and hits call trace due to faulty condition check.

Simplified logic and made station-count-decrement
logic independent of recovery mode. Now, if the
station count is zero, we just return or else we
proceed further to decrement the counter with no
WARN_ON as counter ensured to be zero already.

Fixes: de1456a659f6 ("wifi: ath12k: Add support for Mode 1 recovery")
Change-Id: I6afa7bbbd4f19982da6ab00587fff6425f1a00bb
Signed-off-by: Chetan Kumar Sanga <quic_csanga@quicinc.com>