wifi: ath12k: Fix BUG: scheduling while atomic during SAWF traffic
When SAWF DL(downlink) and UL(Uplink) traffics are started, we will
receive callbacks from ECM module to ath12k module to return
MSDUQ ID and send the UL configurations to target respectively.
However these callbacks source seems to be from interrupt(softirq)
context, which leads to an issue, when mutex lock which is used
in ath12k goes for scheduling.

To address this issue in UL traffic case, since no return value is
needed to be returned to ECM, make use of a list and queue to send the
UL configurations to target. So now "sawf_svc_lock", will be used
to protect this list along with the service class database.
In DL traffic case, since a return value needs to be sent to ECM
module, change to use spin_lock_bh instead of mutex lock
since the service class database can be accessed during interrupt
context and also in process context. This process context access will
be from userspace. Also make changes to optimize the use of locks.

Change-Id: Ia2bf6821bc215a1729a410ea66e3af20af050385
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>