ath12k: Drop packets of slave interface in management work
During multiclient streaming scenario with direct switch
enabled, sometimes data packets of bonded slave interface are
routed to management work incorrectly.

Because of this, when management work tries to handle this data
frame skb_over_panic is seen as we dont have the size to write the data.

Ideally, these data packets should not be routed to management work.
In any case, a packet from slave interface should not be
handled in management work and therefore, it should be dropped.

Change-Id: Ice42674a1f1660ab3b8f244edb25c24a406285e0
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>