wifi: ath12k: fix channel switch between two radios
Currently if channel is changed from one underlying radio to another, it is
not checked and hence driver proceeds to send wrong frequency info of the
channel to the wrong underlying hardware which ultimately leads to firmware
assert.

Fix this issue by not allowing channel to switch between two radios and
returning operation not supported to upper layer.

Change-Id: I58b2e94a10cfae9e21cf2b83da224dea23b7314e
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>