ath12k: fix incorrect type cast of arsta in set key
During SSR mode1 recovery, driver would try to flush the
cached data via ath12k_mac_vif_cache_flush() api once
vdev is successfully created. While flusing the key_conf,
we are facing kernel panic with invalid address access.

This is because, during flush, there is a sta entry in key_conf
which is of ieee80211_sta type and that is sent to ath12k_mac_set_key()
as a parameter which is of type ath12k_link_sta.

Because of this incorrect type being passed, the values that is fetched
are improper and leading to crash.

Hence, add changes to pass proper arsta for ath12k_mac_set_key() api
from ath12k_mac_vif_cache_flush().

Fixes: I1a3beeb7f1b ("ath12k: Fix failed to remove key issue")

Change-Id: Id7171331b5329290e2ba2822d2f1202dbc440489
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>