wifi: ath12k: fix disabling channels based on boot time regulatory capability
Currently, during setting up channels, channels are marked as disabled if
it does not fall into the frequency range. The frequency range is calculated
based on the minimum of target advertised capability and the boot time
regulatory event from target to host. This leads to issue in enabling the
channels back when later regulatory is changed and the frequency range is
expanded.

For example, target supports minimum 2401 MHz and maximum 2484 MHz and the
same is advertised in target capabilities. Now by default, currently boot
time country is US so reg rules will be sent from target to host and for
2 GHz, reg rule frequency range is 2402 - 2472 MHz. Hence, the final
frequency range would be taken as 2402 - 2472 MHz. During, setting up
channels, channel 14 in 2 GHz will be marked as disabled now. This will
make channel 14's original flags to carry disabled flag.

Later, when regulatory is updated to cfg80211 by host, cfg80211 process the
reg rules and marks channel 12 and 13 as well as disabled. But this marking
does not make the original flags as disabled.

Now if regulatory is changed to Japan (JP), target sends reg rules for 2 GHz
as 2402 - 2482 MHz. When host updates it to cfg80211, cfg80211 tries to apply
the rule in the 2 GHz channel list. Following happens -
  * Channel 12 and Channel 13 falls in the reg rules hence it is marked as
    enabled. At the same time, their original flag is also copied. Since the
    original flag doesn't have disabled flag, Channel 12 and Channel 13 gets
    enabled ultimately.
  * Channel 14 also lies in the reg rule, hence that is also marked enabled.
    When copying the original flags, Channel 14 has disabled flag, so disabled
    flags gets copied. And hence, Channel 14 still remains disabled.

The above is wrong since in JP country, Channel 14 is allowed in DSS modes.

This happened because, initially Channel 14 had disabled set in original flags
and that happened cause while setting channels, regulatory rule from target to
host was also considered. However, there is no need of it. Only target
capability is sufficient.

Fix the above issue by only marking channels as disabled during setting up
channels based on the target capabilities.

Fixes: e810d1263e8e ("ath12k: supporting 5.9G band with 160Mhz")
Change-Id: Icc8e62da9d6590a1eac55f559491fc3d079217ed
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
(cherry picked from commit 6c08f64d2f1badf9b8b5f5614dd74af9b548ecd7)