ath12k: enable ce manual poll for qcn9274
If there is no buffer to send packet via host->target, then check if
interrupts are not processed from that CE for last 500ms. If so,
reap the ring to allow new buffers to be enqueued. These changes
were added as part of hw parameter support_ce_manual_poll.

There is a case where intermittently completion interrupts are
not received from CE3 and hence, enable the support_ce_manual_poll
for qcn9274 to come out of this issue scenario.

Change-Id: Ic12496af296ca9eef5b8866b9250847552f06d9b
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>