ath12k: fill ath12k_coredump_ram_info in case of host crash in recovery support
In case of host crash with recovery enabled,
ath12k_coredump_build_inline() won't be called as it is
already in recovery and full dump will be collected from
uboot but the contents of ath12k_coredump_ram_info won't
be filled as it is in recovery path this will lead to
failure in firmware dump extraction while parsing.

So, fill the contents of ath12k_coredump_ram_info even if
recovery support is there and if we are in panic path.

Change-Id: Ic2563be932e25a6ebc19c9dd90472d45f8f61743
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>