ath12k: enable SSR Mode0 failsafe mode by default
When SSR Mode0 fails to reconfigure the chip due to a host or a FW
issue, then the chip will be in neither recoverable nor re-assertable
state.

SSR Mode0 failsafe mechanism to issue a BUG_ON during above state is
there as a module param but not enabled by default.

Enable it by default to assist in debugging FW assert/crash during
SSR Mode0 recovery by crashing the system to collect coredump.

Change-Id: Ifaa2b6f52dbb9bf8df9f033005d1a3b597e7cd7d
Signed-off-by: Sathishkumar Muruganandam <quic_murugana@quicinc.com>