wifi: ath12k: Fix beacon rx for sta associated to non Tx AP
Currently, When a station is associated to a non Tx AP, beacons are not
received frequently from FW. This results in missing events via beacons
such as channel switch leading to sta vif not switching to new channel as
the AP and causing kickout event from FW eventually.

This was due to missing configuration of non Tx info during STA vdev up.
Fill the info such as Tx bssid, profile index and count for receiving
beacons from the Tx AP properly.

Change-Id: Icf48fdc36c52ee45afce37dd5fc08319af8b2d86
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>