wifi: ath11k: ath12k: Add support to send the QoS Null Data frame through exception path
When we try to send QoS NULL Data frame in Ethernet encap type, it modified
as QoS Data frame with TID 0 (encryption enabled one if security enabled).
But expectation is, it should be send as open type frame with TID 7 since
its a QoS NULL data frame.

Added this frame under exception route to bypass TCL with the help of FW.

In MSDU EXT descriptor enqueue, TCL not considered the packet offset. So if we
pack HTT META then it misbehave due to misconception of HTT META as packet data.
So avoid the HTT META data pack for the MSDU EXT descriptor.

TID information is not fetched correctly from the packet header due to
invalid packet offset configuration in the TCL command descriptor, When
the exception tx path choosed. In exception tx packet, HTT meta populated
before the packet payload. So fix this issue by the correct packet offset
configuration in the TCL command descriptor.

Fixes: Ib7828122f43e393bf ("ath12k: Extend Tx optimization changes to mesh path and further optimize")

Change-Id: Ie5ecba48d656bb50b8f105ece38577d3f600a403
Signed-off-by: Karthikeyan Periyasamy <quic_periyasa@quicinc.com>